requirements.txt
google-generativeai
python-dotenv
docx2txt